Restructure repo.
