+++ src/feature_engineering/image_generator_using_simhash.py
@@ -1,6 +1,6 @@
 import numpy as np
 from PIL import Image
-import hash_generator
+import src.hashing.hash_generator as hash_generator
 import os
 
 class SimHashGenerator:
